♻️ Refactor:component modularization
