+++ src/assets/svg/index.ts
@@ -51,3 +51,4 @@ export { default as NewPageIcon } from './NewPageIcon';
 export { default as TrashcanIcon } from './TrashcanIcon';
 export { default as SettingIcon } from './SettingIcon';
 export { default as CircleDefaultProfile } from './CircleDefaultProfile';
+export { default as BookBenIcon } from './BookBenIcon';
