+++ src/components/StuInfo/atoms/Gender/style.ts
@@ -1,0 +1,5 @@
+import styled from "@emotion/styled";
+
+export const Gender = styled.p`
+  flex-basis: 1em;
+`;
