+++ src/pages/studsDetail.tsx
@@ -9,12 +9,9 @@ import {
   StuInfoTemplate,
 } from 'components/StuInfo/template/style';
 import UseThemeEffect from 'hooks/useThemeEffect';
-import { GetServerSideProps, NextPage } from 'next';
+import { NextPage } from 'next';
 import { SWRConfig } from 'swr';
 import { StuInfoType } from 'types/components/StuInfoPage';
-import { apiClient } from 'utils/Libs/apiClient';
-import { getToken } from 'utils/Libs/getToken';
-import { StuInfoController } from 'utils/Libs/requestUrls';
 
 const StudsDetail: NextPage<{
   fallback: Record<string, StuInfoType[]>;
@@ -37,32 +34,6 @@ const StudsDetail: NextPage<{
       </SWRConfig>
     </>
   );
-};
-
-export const getServerSideProps: GetServerSideProps = async (ctx) => {
-  const { Authorization } = await getToken(ctx);
-
-  try {
-    const { data: stuInfo } = await apiClient.get(StuInfoController.stuInfo, {
-      headers: { Authorization },
-    });
-    const { data: searchStuInfo } = await apiClient.get(
-      StuInfoController.stuInfo,
-      {
-        headers: { Authorization },
-      }
-    );
-    return {
-      props: {
-        fallback: {
-          [StuInfoController.stuInfo]: stuInfo,
-          [StuInfoController.searchStuInfo]: searchStuInfo,
-        },
-      },
-    };
-  } catch (e) {
-    return { props: {} };
-  }
 };
 
 export default StudsDetail;
